+++ Engine/Source/Runtime/Windows/D3D11RHI/Private/D3D11ConstantBuffer.cpp
@@ -16,7 +16,7 @@
 const uint32 GConstantBufferSizes[MAX_CONSTANT_BUFFER_SLOTS] = 
 {
 	// CBs must be a multiple of 16
-	Align(MAX_GLOBAL_CONSTANT_BUFFER_SIZE, 16),
+	(uint32)Align(MAX_GLOBAL_CONSTANT_BUFFER_SIZE, 16),
 };
 
 // New circular buffer system for faster constant uploads.  Avoids CopyResource and speeds things up considerably
